hashaul-main/HashHaul/backend.v2/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
marshmallow==3.20.1
PyMySQL==1.1.0
cryptography==41.0.4